requirements.txt
anyio==3.3.4
argon2-cffi==21.1.0
asgiref==3.4.1
black==21.10b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.70.0
fastapi_csrf_protect==0.2.1
itsdangerous==2.0.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
regex==2021.11.2
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
tomli==1.2.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
